+++ src/constants.ts
@@ -1,5 +1,7 @@
-export const groupsJsonFileName = 'groups.json';
+export const projectsJsonFileName = 'projects.json';
+export const projectsJsoncFileName = 'projects.jsonc';
 export const proDashFolderName = '.prodash';
 export const scriptsJsonFileName = 'scripts.json';
+export const scriptsJsoncFileName = 'scripts.jsonc';
 export const projectDescriptionFileName = 'description.$$$';
-
+export const projectLongDescriptionFileName = 'longdescription.$$$';